work

+++ app/Http/Controllers/API/FieldController.php
@@ -91,9 +91,13 @@ class FieldController extends Controller
      */
     public function update(Request $request, Field $field)
     {
-        // $this->authorize('fields.update');
+        $this->authorize('fields.update');
 
-        $attributes = $request->validate($this->rules);
+        if ($field->locked) {
+            $attributes = $request->except('handle')->validate($this->rules);
+        } else {
+            $attributes = $request->validate($this->rules);
+        }
 
         $field->update($attributes);
 
@@ -102,7 +106,7 @@ class FieldController extends Controller
 
     public function reorder(Request $request)
     {
-        // $this->authorize('fields.update');
+        $this->authorize('fields.update');
 
         $fields = $request->get('fields');
 
@@ -124,8 +128,10 @@ class FieldController extends Controller
      */
     public function destroy(Field $field)
     {
-        // $this->authorize('fields.delete');
+        $this->authorize('fields.delete');
 
-        $field->delete();
+        if (! $field->locked) {
+            $field->delete();
+        }
     }
 }
